pyproject.toml
[tool.poetry]
name = "scoap3"
version = "0.1.0"
description = ""
authors = ["CERN SIS-TS <cern-sis-ts-admins@cern.ch>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "~3.11"
pytz = "^2023.3"
python-slugify = "^8.0.1"
Pillow = "^9.5.0"
argon2-cffi = "^21.3.0"
whitenoise = "^6.4.0"
redis = "^4.5.5"
hiredis = "^2.2.3"
celery = "^5.3.1"
django-celery-beat = "^2.5.0"
flower = "^2.0.0"
Django = "^4.2"
django-environ = "^0.10.0"
django-model-utils = "^4.3.1"
django-allauth = "^0.54.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-redis = "^5.3.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
drf-spectacular = "^0.26.1"
django-webpack-loader = "^2.0.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.6"
sentry-sdk = "^1.19.1"
django-storages = "^1.13.2"
django-anymail = "^10.0"
django-elasticsearch-dsl-drf = {git = "https://github.com/cern-sis/django-elasticsearch-dsl-drf.git", rev = "opensearch"}
django-opensearch-dsl = "^0.5.1"
elasticsearch = "6.8.2"
pycountry = "^22.3.5"
boto3 = "^1.26.161"
backoff = "^2.2.1"

[tool.poetry.dev-dependencies]
Werkzeug = {extras = ["watchdog"], version = "^2.3.4"}
ipdb = "^0.13.13"
psycopg2 = "^2.9.6"
watchfiles = "^0.19.0"
mypy = "^1.3.0"
django-stubs = "^4.2.1"
pytest = "^7.3.0"
pytest-sugar = "^0.9.7"
djangorestframework-stubs = "^3.14.0"
Sphinx = "^7.0.1"
sphinx-autobuild = "^2021.3.14"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"
coverage = "^7.2.7"
black = "^23.3.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
pre-commit = "^3.3.3"
factory-boy = "^3.2.1"
django-debug-toolbar = "^4.1.0"
django-extensions = "^3.2.3"
django-coverage-plugin = "^3.0.0"
pytest-django = "^4.5.2"

[tool.isort]
profile = "black"
multi_line_output = 3
atomic = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--ds=config.settings.test", "--reuse-db"]
python_files = ["test_*.py", "*_test.py"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
    ]
